Add hirescoarsekmax, hirescoarsepeaks methods
These methods read in a single hi-res time signal as a Vector, reshape
it into an Matrix of size (nfpc, ncoarse=nfine÷nfpc).  hirescoarsekmax
then computes the kurtosis of each coarse channel, returning a Vector of
jurtosis values, one per coarse channel.  hirescoarsepeaks does a
`findmax` on each coarse channel and returns a Vector of peak values per
coarse channel and a Vector of peak indices (as a Vector of
CartesianIndex values).